+++ fv3config/fv3run/_native.py
@@ -69,6 +69,9 @@ def run_native(
             a configuration dictionary
         outdir (str): location to copy the resulting run directory
         runfile (str, optional): Python model script to use in place of the default.
+        capture_output (bool, optional): If true, then the stderr and stdout
+            streams will be redirected to the files `outdir/stderr.log` and `outdir/stdout.log`
+            respectively.
     """
     _set_stacksize_unlimited()
     with _temporary_directory(outdir) as localdir:
@@ -83,13 +86,15 @@ def run_native(
             filesystem.get_file(
                 runfile, os.path.join(localdir, os.path.basename(runfile))
             )
-        with _log_exceptions(localdir):
+        with _output_stream_context(localdir, capture_output) as (stdout, stderr):
             n_processes = get_n_processes(config_dict)
             _run_experiment(
                 localdir,
                 n_processes,
                 runfile=runfile,
                 mpi_flags=_add_oversubscribe_if_necessary(MPI_FLAGS, n_processes),
+                stdout=stdout,
+                stderr=stderr,
             )
 
 
@@ -133,18 +138,36 @@ def _temporary_directory(outdir):
         yield outdir
 
 
-@contextlib.contextmanager
-def _log_exceptions(localdir):
-    logger.info("running experiment")
+def _captured_output_context(localdir):
+    out_filename = os.path.join(localdir, STDOUT_FILENAME)
+    err_filename = os.path.join(localdir, STDERR_FILENAME)
+    with open(out_filename, "wb") as out_file, open(err_filename, "wb") as err_file:
+        try:
+            yield out_file, err_file
+        except subprocess.CalledProcessError as e:
+            logger.critical(
+                "Experiment failed. " "Check %s and %s for logs.",
+                STDOUT_FILENAME,
+                STDERR_FILENAME,
+            )
+            raise e
+
+
+def _uncaptured_output_context(localdir):
     try:
-        yield
+        yield sys.stdout, sys.stderr
     except subprocess.CalledProcessError as e:
-        logger.critical(
-            "Experiment failed. " "Check %s and %s for logs.",
-            STDOUT_FILENAME,
-            STDERR_FILENAME,
-        )
+        logger.critical("Experiment failed")
         raise e
+
+
+@contextlib.contextmanager
+def _output_stream_context(localdir: str, capture_output: bool):
+    logger.info("running experiment")
+    if capture_output:
+        yield from _captured_output_context(localdir)
+    else:
+        yield from _uncaptured_output_context(localdir)
 
 
 def _get_python_command(runfile):
@@ -158,21 +181,20 @@ def _get_python_command(runfile):
     return python_args
 
 
-def _run_experiment(dirname, n_processes, runfile, mpi_flags=None):
+def _run_experiment(
+    dirname, n_processes, runfile, mpi_flags=None, stdout=None, stderr=None
+):
     if mpi_flags is None:
         mpi_flags = []
 
     python_command = _get_python_command(runfile)
-    out_filename = os.path.join(dirname, STDOUT_FILENAME)
-    err_filename = os.path.join(dirname, STDERR_FILENAME)
-    with open(out_filename, "wb") as out_file, open(err_filename, "wb") as err_file:
-        logger.info("Running experiment in %s", dirname)
-        subprocess.check_call(
-            ["mpirun", "-n", str(n_processes)] + mpi_flags + python_command,
-            cwd=dirname,
-            stdout=out_file,
-            stderr=err_file,
-        )
+    logger.info("Running experiment in %s", dirname)
+    subprocess.check_call(
+        ["mpirun", "-n", str(n_processes)] + mpi_flags + python_command,
+        cwd=dirname,
+        stdout=stderr,
+        stderr=stdout,
+    )
 
 
 def _get_config_dict_and_write(config_dict_or_location, config_out_filename):