+++ HISTORY.rst
@@ -11,6 +11,7 @@ Major changes:
 
 Minor changes:
 ~~~~~~~~~~~~~
+- Control whether outputs are logged to console or not in `run_kubernetes`, `run_native`, and `run_docker`.
 - Added the flag ``--mca btl_vader_single_copy_mechanism none to mpirun in fv3run`` to mpirun in fv3run
 - Add ReadTheDocs configuration file
 - Do not require output dir and fv3config to be remote in ``run_kubernetes``
@@ -18,6 +19,8 @@ Minor changes:
 
 Breaking changes
 ~~~~~~~~~~~~~~~~
+- Print stderr and stdout to the console by default when using fv3run. Use the
+  `--capture-output` command-line flag to enable the previous behavior.
 - Refactored run_kubernetes and run_docker to call run_native via a new API serializing
   their args/kwargs as json strings. The
   fv3config version in a docker image must be greater than or equal inside a